task1_data_acquisition_preprocessing/ff++ (c23)/requirements.txt
opencv-python
os
pyyaml
time